requirements.txt
psycopg[binary]==3.0.16
python-dotenv
